Refactor: Splitting useTestPlatform into smaller hooks
I've broken down the large and complex useTestPlatform hook into several smaller, more focused hooks:

- useTimer: Manages all timer-related logic.
- useWebcam: Manages all logic related to the webcam and face detection.
- useCodeExecution: Manages the execution of your code and the evaluation of test results.

This change improves readability, maintainability, and separation of concerns within the codebase. It also makes each hook easier to test individually.
